network connect disconnect on non-running containers
a container can connect and disconnet to networks even when not in a
running state.
