+++ test/e2e/network_test.go
@@ -347,159 +347,6 @@ var _ = Describe("Podman network", func() {
 		Expect(c3.ExitCode()).To(BeZero())
 	})
 
-	It("bad network name in disconnect should result in error", func() {
-		SkipIfRootless("network connect and disconnect are only rootfull")
-		dis := podmanTest.Podman([]string{"network", "disconnect", "foobar", "test"})
-		dis.WaitWithDefaultTimeout()
-		Expect(dis.ExitCode()).ToNot(BeZero())
-
-	})
-
-	It("bad container name in network disconnect should result in error", func() {
-		SkipIfRootless("network connect and disconnect are only rootfull")
-		netName := "aliasTest" + stringid.GenerateNonCryptoID()
-		session := podmanTest.Podman([]string{"network", "create", netName})
-		session.WaitWithDefaultTimeout()
-		Expect(session.ExitCode()).To(BeZero())
-		defer podmanTest.removeCNINetwork(netName)
-
-		dis := podmanTest.Podman([]string{"network", "disconnect", netName, "foobar"})
-		dis.WaitWithDefaultTimeout()
-		Expect(dis.ExitCode()).ToNot(BeZero())
-
-	})
-
-	It("podman network disconnect with invalid container state should result in error", func() {
-		SkipIfRootless("network connect and disconnect are only rootfull")
-		netName := "aliasTest" + stringid.GenerateNonCryptoID()
-		session := podmanTest.Podman([]string{"network", "create", netName})
-		session.WaitWithDefaultTimeout()
-		Expect(session.ExitCode()).To(BeZero())
-		defer podmanTest.removeCNINetwork(netName)
-
-		ctr := podmanTest.Podman([]string{"create", "--name", "test", "--network", netName, ALPINE, "top"})
-		ctr.WaitWithDefaultTimeout()
-		Expect(ctr.ExitCode()).To(BeZero())
-
-		dis := podmanTest.Podman([]string{"network", "disconnect", netName, "test"})
-		dis.WaitWithDefaultTimeout()
-		Expect(dis.ExitCode()).ToNot(BeZero())
-	})
-
-	It("podman network disconnect", func() {
-		SkipIfRootless("network connect and disconnect are only rootfull")
-		netName := "aliasTest" + stringid.GenerateNonCryptoID()
-		session := podmanTest.Podman([]string{"network", "create", netName})
-		session.WaitWithDefaultTimeout()
-		Expect(session.ExitCode()).To(BeZero())
-		defer podmanTest.removeCNINetwork(netName)
-
-		ctr := podmanTest.Podman([]string{"run", "-dt", "--name", "test", "--network", netName, ALPINE, "top"})
-		ctr.WaitWithDefaultTimeout()
-		Expect(ctr.ExitCode()).To(BeZero())
-
-		exec := podmanTest.Podman([]string{"exec", "-it", "test", "ip", "addr", "show", "eth0"})
-		exec.WaitWithDefaultTimeout()
-		Expect(exec.ExitCode()).To(BeZero())
-
-		dis := podmanTest.Podman([]string{"network", "disconnect", netName, "test"})
-		dis.WaitWithDefaultTimeout()
-		Expect(dis.ExitCode()).To(BeZero())
-
-		exec = podmanTest.Podman([]string{"exec", "-it", "test", "ip", "addr", "show", "eth0"})
-		exec.WaitWithDefaultTimeout()
-		Expect(exec.ExitCode()).ToNot(BeZero())
-	})
-
-	It("bad network name in connect should result in error", func() {
-		SkipIfRootless("network connect and disconnect are only rootfull")
-		dis := podmanTest.Podman([]string{"network", "connect", "foobar", "test"})
-		dis.WaitWithDefaultTimeout()
-		Expect(dis.ExitCode()).ToNot(BeZero())
-
-	})
-
-	It("bad container name in network connect should result in error", func() {
-		SkipIfRootless("network connect and disconnect are only rootfull")
-		netName := "aliasTest" + stringid.GenerateNonCryptoID()
-		session := podmanTest.Podman([]string{"network", "create", netName})
-		session.WaitWithDefaultTimeout()
-		Expect(session.ExitCode()).To(BeZero())
-		defer podmanTest.removeCNINetwork(netName)
-
-		dis := podmanTest.Podman([]string{"network", "connect", netName, "foobar"})
-		dis.WaitWithDefaultTimeout()
-		Expect(dis.ExitCode()).ToNot(BeZero())
-
-	})
-
-	It("podman connect on a container that already is connected to the network should error", func() {
-		SkipIfRootless("network connect and disconnect are only rootfull")
-		netName := "aliasTest" + stringid.GenerateNonCryptoID()
-		session := podmanTest.Podman([]string{"network", "create", netName})
-		session.WaitWithDefaultTimeout()
-		Expect(session.ExitCode()).To(BeZero())
-		defer podmanTest.removeCNINetwork(netName)
-
-		ctr := podmanTest.Podman([]string{"create", "--name", "test", "--network", netName, ALPINE, "top"})
-		ctr.WaitWithDefaultTimeout()
-		Expect(ctr.ExitCode()).To(BeZero())
-
-		con := podmanTest.Podman([]string{"network", "connect", netName, "test"})
-		con.WaitWithDefaultTimeout()
-		Expect(con.ExitCode()).ToNot(BeZero())
-	})
-
-	It("podman network connect with invalid container state should result in error", func() {
-		SkipIfRootless("network connect and disconnect are only rootfull")
-		netName := "aliasTest" + stringid.GenerateNonCryptoID()
-		session := podmanTest.Podman([]string{"network", "create", netName})
-		session.WaitWithDefaultTimeout()
-		Expect(session.ExitCode()).To(BeZero())
-		defer podmanTest.removeCNINetwork(netName)
-
-		ctr := podmanTest.Podman([]string{"create", "--name", "test", "--network", netName, ALPINE, "top"})
-		ctr.WaitWithDefaultTimeout()
-		Expect(ctr.ExitCode()).To(BeZero())
-
-		dis := podmanTest.Podman([]string{"network", "connect", netName, "test"})
-		dis.WaitWithDefaultTimeout()
-		Expect(dis.ExitCode()).ToNot(BeZero())
-	})
-
-	It("podman network connect", func() {
-		SkipIfRemote("This requires a pending PR to be merged before it will work")
-		SkipIfRootless("network connect and disconnect are only rootfull")
-		netName := "aliasTest" + stringid.GenerateNonCryptoID()
-		session := podmanTest.Podman([]string{"network", "create", netName})
-		session.WaitWithDefaultTimeout()
-		Expect(session.ExitCode()).To(BeZero())
-		defer podmanTest.removeCNINetwork(netName)
-
-		ctr := podmanTest.Podman([]string{"run", "-dt", "--name", "test", "--network", netName, ALPINE, "top"})
-		ctr.WaitWithDefaultTimeout()
-		Expect(ctr.ExitCode()).To(BeZero())
-
-		exec := podmanTest.Podman([]string{"exec", "-it", "test", "ip", "addr", "show", "eth0"})
-		exec.WaitWithDefaultTimeout()
-		Expect(exec.ExitCode()).To(BeZero())
-
-		// Create a second network
-		newNetName := "aliasTest" + stringid.GenerateNonCryptoID()
-		session = podmanTest.Podman([]string{"network", "create", newNetName})
-		session.WaitWithDefaultTimeout()
-		Expect(session.ExitCode()).To(BeZero())
-		defer podmanTest.removeCNINetwork(newNetName)
-
-		connect := podmanTest.Podman([]string{"network", "connect", newNetName, "test"})
-		connect.WaitWithDefaultTimeout()
-		Expect(connect.ExitCode()).To(BeZero())
-
-		exec = podmanTest.Podman([]string{"exec", "-it", "test", "ip", "addr", "show", "eth1"})
-		exec.WaitWithDefaultTimeout()
-		Expect(exec.ExitCode()).To(BeZero())
-	})
-
 	It("podman network create/remove macvlan", func() {
 		net := "macvlan" + stringid.GenerateNonCryptoID()
 		nc := podmanTest.Podman([]string{"network", "create", "--macvlan", "lo", net})
